Split OAI client into cohesive files and add audit write mirroring.
Refs: #7. Addresses admin feedback (thejhh): extract helpers to smaller files.
