+++ internal/oai/client.go
@@ -1,71 +1,23 @@
 package oai
 
 import (
-	"bufio"
-	"bytes"
-	"context"
-	"crypto/rand"
-	"encoding/hex"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	mathrand "math/rand"
-	"net"
-	"net/http"
-	"net/http/httptrace"
-	"os"
-	"path/filepath"
-	"strings"
-	"time"
+    "bytes"
+    "context"
+    "encoding/json"
+    "errors"
+    "fmt"
+    "io"
+    "net/http"
+    "net/http/httptrace"
+    "strings"
+    "time"
 )
 
 type Client struct {
-	baseURL    string
-	apiKey     string
-	httpClient *http.Client
-	retry      RetryPolicy
-}
-
-// audit context keys are unexported to avoid collisions. Use helper to set.
-type auditCtxKey string
-
-const (
-	auditCtxKeyStage auditCtxKey = "audit_stage"
-)
-
-// WithAuditStage returns a child context that carries an audit stage label
-// (e.g., "prep") that will be included in HTTP audit entries.
-func WithAuditStage(parent context.Context, stage string) context.Context {
-	stage = strings.TrimSpace(stage)
-	if stage == "" {
-		return parent
-	}
-	return context.WithValue(parent, auditCtxKeyStage, stage)
-}
-
-func auditStageFromContext(ctx context.Context) string {
-	if ctx == nil {
-		return ""
-	}
-	if v := ctx.Value(auditCtxKeyStage); v != nil {
-		if s, ok := v.(string); ok {
-			return s
-		}
-	}
-	return ""
-}
-
-// RetryPolicy controls HTTP retry behavior for transient failures.
-// MaxRetries specifies the number of retries after the initial attempt.
-// Backoff specifies the base delay between attempts; exponential backoff is applied.
-// JitterFraction specifies the +/- fractional jitter applied to each computed backoff.
-// When Rand is non-nil, it is used to sample jitter for deterministic tests.
-type RetryPolicy struct {
-	MaxRetries     int
-	Backoff        time.Duration
-	JitterFraction float64
-	Rand           *mathrand.Rand
+    baseURL    string
+    apiKey     string
+    httpClient *http.Client
+    retry      RetryPolicy
 }
 
 // NewClient creates a client without retries (single attempt only).
@@ -355,354 +307,3 @@ func (c *Client) StreamChat(ctx context.Context, req ChatCompletionsRequest, onC
 	}
 }
 
-// newLineReader returns a closure that reads one line (terminated by \n) from r each call.
-func newLineReader(r io.Reader) func() (string, error) {
-	br := bufio.NewReader(r)
-	return func() (string, error) {
-		b, err := br.ReadBytes('\n')
-		if err != nil {
-			return "", err
-		}
-		return string(b), nil
-	}
-}
-
-func truncate(s string, n int) string {
-	if len(s) <= n {
-		return s
-	}
-	return s[:n]
-}
-
-// isRetryableError returns true for transient network/timeouts.
-func isRetryableError(err error) bool {
-	if err == nil {
-		return false
-	}
-	// Context deadline exceeded from client timeout
-	if errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled) {
-		return true
-	}
-	var ne net.Error
-	if errors.As(err, &ne) {
-		if ne.Timeout() { // ne.Temporary is deprecated; avoid
-			return true
-		}
-	}
-	// *url.Error often wraps retryable errors; fall back to string contains of "timeout"
-	s := strings.ToLower(err.Error())
-	return strings.Contains(s, "timeout")
-}
-
-// sleepBackoff retained for backward compatibility; not used.
-// Deprecated: use backoffDuration + sleepFor instead.
-// func sleepBackoff(base time.Duration, attempt int) { time.Sleep(backoffDuration(base, attempt)) }
-
-// backoffDuration returns the duration that sleepBackoff would sleep for a given attempt.
-func backoffDuration(base time.Duration, attempt int) time.Duration {
-	if base <= 0 {
-		base = 200 * time.Millisecond
-	}
-	d := base << attempt
-	if d > 2*time.Second {
-		d = 2 * time.Second
-	}
-	return d
-}
-
-// backoffWithJitter returns an exponential backoff adjusted by +/- jitter fraction.
-// When jitterFraction <= 0, this falls back to backoffDuration. When r is nil,
-// a time-seeded RNG is used for production randomness.
-func backoffWithJitter(base time.Duration, attempt int, jitterFraction float64, r *mathrand.Rand) time.Duration {
-	d := backoffDuration(base, attempt)
-	if jitterFraction <= 0 {
-		return d
-	}
-	if jitterFraction > 0.9 { // prevent extreme factors
-		jitterFraction = 0.9
-	}
-	if r == nil {
-		// Seed with current time for production; tests can pass a custom Rand
-		r = mathrand.New(mathrand.NewSource(time.Now().UnixNano()))
-	}
-	// factor in [1 - f, 1 + f]
-	minF := 1.0 - jitterFraction
-	maxF := 1.0 + jitterFraction
-	factor := minF + r.Float64()*(maxF-minF)
-	// Guard against rounding to zero
-	jittered := time.Duration(float64(d) * factor)
-	if jittered < time.Millisecond {
-		return time.Millisecond
-	}
-	return jittered
-}
-
-// retryAfterDuration parses the Retry-After header which may be seconds or HTTP-date.
-// Returns (duration, true) when valid; otherwise (0, false).
-func retryAfterDuration(h string, now time.Time) (time.Duration, bool) {
-	h = strings.TrimSpace(h)
-	if h == "" {
-		return 0, false
-	}
-	// Try integer seconds first
-	if secs, err := time.ParseDuration(h + "s"); err == nil {
-		if secs > 0 {
-			return secs, true
-		}
-	}
-	// Try HTTP-date formats per RFC 7231 (use http.TimeFormat)
-	if t, err := time.Parse(http.TimeFormat, h); err == nil {
-		if t.After(now) {
-			return t.Sub(now), true
-		}
-	}
-	return 0, false
-}
-
-// sleepFor sleeps for the provided duration; extracted for testability.
-// sleepFunc allows tests to intercept sleeps deterministically.
-var sleepFunc = sleepFor
-
-func sleepFor(d time.Duration) {
-	if d <= 0 {
-		return
-	}
-	time.Sleep(d)
-}
-
-// generateIdempotencyKey returns a random hex string suitable for Idempotency-Key.
-func generateIdempotencyKey() string {
-	var b [16]byte
-	if _, err := rand.Read(b[:]); err != nil {
-		// Fallback to timestamp-based key if crypto/rand fails; extremely unlikely
-		return fmt.Sprintf("goagent-%d", time.Now().UnixNano())
-	}
-	return "goagent-" + hex.EncodeToString(b[:])
-}
-
-// logHTTPAttempt appends an NDJSON line describing an HTTP attempt and planned backoff.
-func logHTTPAttempt(stage, idemKey string, attempt, maxAttempts, status int, backoffMs int64, endpoint, errStr string) {
-	type audit struct {
-		TS             string `json:"ts"`
-		Event          string `json:"event"`
-		Stage          string `json:"stage,omitempty"`
-		IdempotencyKey string `json:"idempotency_key,omitempty"`
-		Attempt        int    `json:"attempt"`
-		Max            int    `json:"max"`
-		Status         int    `json:"status"`
-		BackoffMs      int64  `json:"backoffMs"`
-		Endpoint       string `json:"endpoint"`
-		Error          string `json:"error,omitempty"`
-	}
-	entry := audit{
-		TS:             time.Now().UTC().Format(time.RFC3339Nano),
-		Event:          "http_attempt",
-		Stage:          stage,
-		IdempotencyKey: idemKey,
-		Attempt:        attempt,
-		Max:            maxAttempts,
-		Status:         status,
-		BackoffMs:      backoffMs,
-		Endpoint:       endpoint,
-		Error:          truncate(errStr, 500),
-	}
-	if err := appendAuditLog(entry); err != nil {
-		_ = err
-	}
-}
-
-// logHTTPTiming appends detailed HTTP timing metrics to the audit log.
-func logHTTPTiming(stage, idemKey string, attempt int, endpoint string, status int, start time.Time, dnsDur, connDur, tlsDur time.Duration, wroteAt, firstByteAt, end time.Time, cause, hint string) {
-	type timing struct {
-		TS             string `json:"ts"`
-		Event          string `json:"event"`
-		Stage          string `json:"stage,omitempty"`
-		IdempotencyKey string `json:"idempotency_key,omitempty"`
-		Attempt        int    `json:"attempt"`
-		Endpoint       string `json:"endpoint"`
-		Status         int    `json:"status"`
-		DNSMs          int64  `json:"dnsMs"`
-		ConnectMs      int64  `json:"connectMs"`
-		TLSMs          int64  `json:"tlsMs"`
-		WroteMs        int64  `json:"wroteMs"`
-		TTFBMs         int64  `json:"ttfbMs"`
-		ReadMs         int64  `json:"readMs"`
-		TotalMs        int64  `json:"totalMs"`
-		Cause          string `json:"cause"`
-		Hint           string `json:"hint,omitempty"`
-	}
-	var wroteMs, ttfbMs, readMs int64
-	if !wroteAt.IsZero() {
-		wroteMs = wroteAt.Sub(start).Milliseconds()
-	}
-	if !firstByteAt.IsZero() {
-		if !wroteAt.IsZero() && firstByteAt.After(wroteAt) {
-			ttfbMs = firstByteAt.Sub(wroteAt).Milliseconds()
-		} else {
-			ttfbMs = firstByteAt.Sub(start).Milliseconds()
-		}
-		if end.After(firstByteAt) {
-			readMs = end.Sub(firstByteAt).Milliseconds()
-		}
-	}
-	entry := timing{
-		TS:             time.Now().UTC().Format(time.RFC3339Nano),
-		Event:          "http_timing",
-		Stage:          stage,
-		IdempotencyKey: idemKey,
-		Attempt:        attempt,
-		Endpoint:       endpoint,
-		Status:         status,
-		DNSMs:          dnsDur.Milliseconds(),
-		ConnectMs:      connDur.Milliseconds(),
-		TLSMs:          tlsDur.Milliseconds(),
-		WroteMs:        wroteMs,
-		TTFBMs:         ttfbMs,
-		ReadMs:         readMs,
-		TotalMs:        end.Sub(start).Milliseconds(),
-		Cause:          cause,
-		Hint:           hint,
-	}
-	if err := appendAuditLog(entry); err != nil {
-		_ = err
-	}
-}
-
-// LogLengthBackoff emits a structured NDJSON audit entry describing a
-// length_backoff event triggered by finish_reason=="length". Callers should
-// pass the model identifier, the previous and new completion caps, the
-// effective model context window, and the estimated prompt token count.
-func LogLengthBackoff(model string, prevCap, newCap, window, estimatedPromptTokens int) {
-	type audit struct {
-		TS                    string `json:"ts"`
-		Event                 string `json:"event"`
-		Model                 string `json:"model"`
-		PrevCap               int    `json:"prev_cap"`
-		NewCap                int    `json:"new_cap"`
-		Window                int    `json:"window"`
-		EstimatedPromptTokens int    `json:"estimated_prompt_tokens"`
-	}
-	entry := audit{
-		TS:                    time.Now().UTC().Format(time.RFC3339Nano),
-		Event:                 "length_backoff",
-		Model:                 model,
-		PrevCap:               prevCap,
-		NewCap:                newCap,
-		Window:                window,
-		EstimatedPromptTokens: estimatedPromptTokens,
-	}
-	if err := appendAuditLog(entry); err != nil {
-		_ = err
-	}
-}
-
-// emitChatMetaAudit writes a one-line NDJSON entry describing request-level
-// observability fields such as the effective temperature and whether the
-// temperature parameter is included in the payload for the target model.
-func emitChatMetaAudit(req ChatCompletionsRequest) {
-	// Compute effective temperature based on model support and clamp rules.
-	effectiveTemp, supported := EffectiveTemperatureForModel(req.Model, valueOrDefault(req.Temperature, 1.0))
-	type meta struct {
-		TS                   string  `json:"ts"`
-		Event                string  `json:"event"`
-		Model                string  `json:"model"`
-		TemperatureEffective float64 `json:"temperature_effective"`
-		TemperatureInPayload bool    `json:"temperature_in_payload"`
-	}
-	entry := meta{
-		TS:                   time.Now().UTC().Format(time.RFC3339Nano),
-		Event:                "chat_meta",
-		Model:                req.Model,
-		TemperatureEffective: effectiveTemp,
-		TemperatureInPayload: supported && req.Temperature != nil,
-	}
-	if err := appendAuditLog(entry); err != nil {
-		_ = err
-	}
-}
-
-func valueOrDefault(ptr *float64, def float64) float64 {
-	if ptr == nil {
-		return def
-	}
-	return *ptr
-}
-
-// classifyHTTPCause returns a short cause label for audit based on error/context.
-func classifyHTTPCause(ctx context.Context, err error) string {
-	if err == nil {
-		return "success"
-	}
-	if errors.Is(err, context.DeadlineExceeded) || (ctx != nil && ctx.Err() == context.DeadlineExceeded) {
-		return "context_deadline"
-	}
-	s := strings.ToLower(err.Error())
-	switch {
-	case strings.Contains(s, "server closed") || strings.Contains(s, "connection reset") || strings.Contains(s, "broken pipe"):
-		return "server_closed"
-	case strings.Contains(s, "timeout"):
-		return "timeout"
-	default:
-		return "error"
-	}
-}
-
-// userHintForCause returns a short actionable hint for common failure causes.
-func userHintForCause(ctx context.Context, err error) string {
-	if err == nil {
-		return ""
-	}
-	if errors.Is(err, context.DeadlineExceeded) || (ctx != nil && ctx.Err() == context.DeadlineExceeded) || strings.Contains(strings.ToLower(err.Error()), "timeout") {
-		return "increase -http-timeout or reduce prompt/model latency"
-	}
-	return ""
-}
-
-// appendAuditLog writes an NDJSON audit line to .goagent/audit/YYYYMMDD.log (same location used by tool runner).
-func appendAuditLog(entry any) error {
-	b, err := json.Marshal(entry)
-	if err != nil {
-		return err
-	}
-	root := moduleRoot()
-	dir := filepath.Join(root, ".goagent", "audit")
-	if err := os.MkdirAll(dir, 0o755); err != nil {
-		return err
-	}
-	fname := time.Now().UTC().Format("20060102") + ".log"
-	path := filepath.Join(dir, fname)
-	f, err := os.OpenFile(path, os.O_CREATE|os.O_APPEND|os.O_WRONLY, 0o644)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if cerr := f.Close(); cerr != nil {
-			_ = cerr
-		}
-	}()
-	if _, err := f.Write(append(b, '\n')); err != nil {
-		return err
-	}
-	return nil
-}
-
-// moduleRoot walks upward from the current working directory to locate the directory
-// containing go.mod. If none is found, it returns the current working directory.
-func moduleRoot() string {
-	cwd, err := os.Getwd()
-	if err != nil || cwd == "" {
-		return "."
-	}
-	dir := cwd
-	for {
-		if _, err := os.Stat(filepath.Join(dir, "go.mod")); err == nil {
-			return dir
-		}
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			// Reached filesystem root; fallback to original cwd
-			return cwd
-		}
-		dir = parent
-	}
-}
